Added LoadReplacementPass function load/use/cost analysis
It appears to be working, but it will need quite a bit more work before it is
usable. The code definately needs to be cleaned up quite a bit. I am pushing
this because it is not broken and it will be useful for my teammate to see what
is going on.